requirements.txt
```txt
streamlit
pillow
